astro-platform-central/requirements.txt
pandas
apache-airflow-providers-amazon
scikit-learn